+++ src/test/java/io/github/alterioncorp/jpa/fetch/PathParserTest.java
@@ -3,10 +3,12 @@ package io.github.alterioncorp.jpa.fetch;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import io.github.alterioncorp.jpa.fetch.entities.path.QOrganization;
-import io.github.alterioncorp.jpa.fetch.entities.path.QPerson;
-
 class PathParserTest {
+
+	@Test
+	public void testBuildNode_emptySegments_throws() {
+		Assertions.assertThrows(IllegalArgumentException.class, () -> PathParser.buildNode(new String[0]));
+	}
 
 	@Test
 	public void testBuildFromPaths() {
@@ -19,12 +21,5 @@ class PathParserTest {
 		Assertions.assertNotNull(tree.getRoot("11").getChild("21").getChild("31"));
 		Assertions.assertNotNull(tree.getRoot("11").getChild("21").getChild("32"));
 		Assertions.assertNotNull(tree.getRoot("12").getChild("22"));
-	}
-
-	@Test
-	public void testPathToString() {
-		Assertions.assertEquals("organization", PathParser.pathToString(QPerson.person.organization()));
-		Assertions.assertEquals("organization.persons", PathParser.pathToString(QPerson.person.organization().persons));
-		Assertions.assertEquals("persons", PathParser.pathToString(QOrganization.organization.persons));
 	}
 }